+++ interrupt-ringbuffer/src/ringbuffer.c
@@ -1,7 +1,7 @@
 #include <stdbool.h>
 #include "ringbuffer.h"
 
-#define RINGBUFFER_SIZE                          (uint16_t)512
+#define RINGBUFFER_SIZE                               512
 
 static uint8_t ringbuffer[RINGBUFFER_SIZE];
 static uint16_t head;
@@ -13,7 +13,7 @@ static bool ringbuffer_empty(void);
 ringbuffer_status_t ringbuffer_put(uint8_t c) {
   if (!ringbuffer_full()) {
     ringbuffer[head] = c;
-    head = (uint16_t)((head + 1U) & (RINGBUFFER_SIZE - 1U));  
+    head = (head + 1) & (RINGBUFFER_SIZE - 1);  
     return RINGBUFFER_STATUS_OK;
   }
   
@@ -23,7 +23,7 @@ ringbuffer_status_t ringbuffer_put(uint8_t c) {
 ringbuffer_status_t ringbuffer_get(uint8_t *byte) {
   if (!ringbuffer_empty()) {
     *byte = ringbuffer[tail];
-    tail = (uint16_t)((tail + 1U) & (RINGBUFFER_SIZE - 1U));
+    tail = (tail + 1) & (RINGBUFFER_SIZE - 1);
     return RINGBUFFER_STATUS_OK;
   }
   
@@ -31,12 +31,12 @@ ringbuffer_status_t ringbuffer_get(uint8_t *byte) {
 }
 
 void ringbuffer_init(void) {
-  head = 0U;
-  tail = 0U;
+  head = 0;
+  tail = 0;
 }
 
 static bool ringbuffer_full(void) {
-  return (((head + 1U) & (RINGBUFFER_SIZE - 1U)) == tail);
+  return ((head + 1) & (RINGBUFFER_SIZE - 1)) == tail;
 }
 
 static bool ringbuffer_empty(void) {
